Fail when a payment amount below 1.0 is requested
While the CO API happily returns outstanding payments below 1€ the API
for registering a complete payment will fail with anything below 1€.

To avoid starting payments which we can never complete fails early.

This reuses the error-id used in the mono bundle to indicate too small
payment amounts (<=0), but in this case also works.