Extend the health check to also cover a semi broken CO API
In case the tuition fee API is working but it can't connect to its backend
we would get a 500 with something like "Input stream was empty, there is no entity"
for API calls. The version API still worked fine thought.

Add a new health check that makes the tuition fee API talk to its backend, so
we get properly notified if that is broken.