Add an extra check to make sure that we don't register a payment that is too large
In case something went wrong and we are registering a payment that is larger than
the amount owed, then something went wrong and we need to abort.

Since the CO API doesn't error out in such a case we need to manually check for
the current amount and compare.

Also adds notify() tests while at it.