tuition-fee-api: don't hardcode the auth server path
The code was previously assuming that the CO keycloak realm
was located at "co/public/sec/auth/realms/CAMPUSonline". This changes
in the next CO release, so we have to find a way to discover the right path.

This now uses the "co/public/api/environment" api to figure out the auth
server path and, in addition, uses the ".well-known/openid-configuration"
endpoint to figure out the token endpoint.

In allows CO to change the keycloak realm, or even the whole IDP provider
in the future.

Bonus changes:

* To avoid having to mock all those requests, skip the token auth on most
  tests.
* Enable logging for each request, the (non-audit) logger wasn't passed
  to the connection previously, which I noticed while implementing this.